fix type

+++ filehash/sha1/sha1.c
@@ -270,7 +270,7 @@ void blk_SHA1_Final(unsigned char hashout[20], blk_SHA_CTX *ctx)
     put_be32(hashout + i * 4, ctx->H[i]);
 }
 
-void SHA1(unsigned char hashout[20], const void* data, unsigned long len)
+void SHA1(unsigned char hashout[20], const void* data, size_t len)
 {
   blk_SHA_CTX ctx;
   blk_SHA1_Init(&ctx);
